+++ app/lib/folio/console_cell.rb
@@ -4,6 +4,7 @@ class Folio::ConsoleCell < Folio::ApplicationCell
   include Folio::Cell::HtmlSafeFieldsFor
   include Folio::Console::CellsHelper
   include Folio::Console::ReportsHelper
+  include Folio::Console::PreviewUrlFor
 
   delegate :safe_url_for,
            :through_aware_console_url_for,
@@ -25,30 +26,6 @@ class Folio::ConsoleCell < Folio::ApplicationCell
       ActionController::Base.helpers.sanitize(str, tags: [], attributes: [])
     else
       str
-    end
-  end
-
-  def preview_url_for(record)
-    args = {}
-
-    if record.respond_to?(:published?) && token = record.try(:preview_token)
-      args[Folio::Publishable::PREVIEW_PARAM_NAME] = token unless record.published?
-    end
-
-    if record.respond_to?(:locale)
-      args[:locale] = record.locale
-    elsif ::Rails.application.config.folio_console_add_locale_to_preview_links
-      args[:locale] = I18n.locale
-    end
-
-    if args[:locale] && Folio::Current.site.locale.present? && Folio::Current.site.locales.present? && Folio::Current.site.locales.exclude?(args[:locale].to_s)
-      args[:locale] = Folio::Current.site.locale
-    end
-
-    begin
-      I18n.with_locale(args[:locale] || I18n.locale) { url_for([record, args]) }
-    rescue NoMethodError
-      nil
     end
   end
 end