+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Changed
+- console `preview_url_for` - now defined in `Folio::Console::PreviewUrlFor` and expandable via `Rails.application.config.folio_console_preview_url_for_procs`
+- added `:preview` to the console actions default - hide it if URL is not available
+
 ### Fixed
 - `react_ordered_multiselect` can now show button to fix required field